+++ packages/firebase_auth/test/firebase_auth_test.dart
@@ -28,6 +28,27 @@ const String kMockLanguage = 'en';
 final int kMockCreationTimestamp = DateTime(2019, 1, 1).millisecondsSinceEpoch;
 final int kMockLastSignInTimestamp =
     DateTime.now().subtract(const Duration(days: 1)).millisecondsSinceEpoch;
+const Map<String, dynamic> kMockUser = <String, dynamic>{
+  'isAnonymous': true,
+  'isEmailVerified': false,
+  'creationTimestamp': kMockCreationTimestamp,
+  'lastSignInTimestamp': kMockLastSignInTimestamp,
+  'providerData': <Map<String, String>>[
+    <String, String>{
+      'providerId': kMockProviderId,
+      'uid': kMockUid,
+      'displayName': kMockDisplayName,
+      'photoUrl': kMockPhotoUrl,
+      'email': kMockEmail,
+    },
+  ],
+};
+const Map<String, dynamic> kMockAdditionalUserInfo = <String, dynamic>{
+  'isNewUser': false,
+  'username': 'flutterUser',
+  'providerId': 'testProvider',
+  'profile': <String, dynamic>{'foo': 'bar'},
+};
 
 void main() {
   group('$FirebaseAuth', () {
@@ -51,6 +72,8 @@ void main() {
           case "startListeningAuthState":
             return mockHandleId++;
             break;
+          case "currentUser":
+            return kMockUser;
           case "sendLinkToEmail":
           case "sendPasswordResetEmail":
           case "updateEmail":
@@ -66,7 +89,10 @@ void main() {
             return null;
             break;
           default:
-            return mockFirebaseUser();
+            return <String, dynamic>{
+              'user': kMockUser,
+              'additionalUserInfo': kMockAdditionalUserInfo,
+            };
             break;
         }
       });
@@ -88,27 +114,25 @@ void main() {
           kMockLastSignInTimestamp);
     }
 
-    test('currentUser', () async {
-      final FirebaseUser user = await auth.currentUser();
-      verifyUser(user);
+    void verifyAuthResult(AuthResult result) {
+      verifyUser(result.user);
+      final AdditionalUserInfo additionalUserInfo = result.additionalUserInfo;
       expect(
-        log,
-        <Matcher>[
-          isMethodCall('currentUser',
-              arguments: <String, String>{'app': auth.app.name}),
-        ],
-      );
-    });
+          additionalUserInfo.isNewUser, kMockAdditionalUserInfo['isNewUser']);
+      expect(additionalUserInfo.username, kMockAdditionalUserInfo['username']);
+      expect(
+          additionalUserInfo.providerId, kMockAdditionalUserInfo['providerId']);
+      expect(additionalUserInfo.profile, kMockAdditionalUserInfo['profile']);
+    }
 
-    test('signInAnonymously', () async {
-      final FirebaseUser user = await auth.signInAnonymously();
-      verifyUser(user);
+    test('getIdToken', () async {
+      final FirebaseUser user = await auth.currentUser();
       expect(await user.getIdToken(), equals(kMockIdToken));
       expect(await user.getIdToken(refresh: true), equals(kMockIdToken));
       expect(
         log,
         <Matcher>[
-          isMethodCall('signInAnonymously',
+          isMethodCall('currentUser',
               arguments: <String, String>{'app': auth.app.name}),
           isMethodCall(
             'getIdToken',
@@ -121,6 +145,18 @@ void main() {
             'getIdToken',
             arguments: <String, dynamic>{'refresh': true, 'app': auth.app.name},
           ),
+        ],
+      );
+    });
+
+    test('signInAnonymously', () async {
+      final AuthResult result = await auth.signInAnonymously();
+      verifyAuthResult(result);
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall('signInAnonymously',
+              arguments: <String, String>{'app': auth.app.name}),
         ],
       );
     });
@@ -165,10 +201,11 @@ void main() {
     });
 
     test('signInWithEmailAndLink', () async {
-      await auth.signInWithEmailAndLink(
+      final AuthResult result = await auth.signInWithEmailAndLink(
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -182,11 +219,11 @@ void main() {
     });
 
     test('createUserWithEmailAndPassword', () async {
-      final FirebaseUser user = await auth.createUserWithEmailAndPassword(
+      final AuthResult result = await auth.createUserWithEmailAndPassword(
         email: kMockEmail,
         password: kMockPassword,
       );
-      verifyUser(user);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -226,9 +263,9 @@ void main() {
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -258,8 +295,8 @@ void main() {
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -309,9 +346,9 @@ void main() {
         authToken: kMockIdToken,
         authTokenSecret: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -341,8 +378,8 @@ void main() {
         authToken: kMockIdToken,
         authTokenSecret: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -365,9 +402,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -395,8 +432,8 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -419,9 +456,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -451,8 +488,8 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -476,7 +513,8 @@ void main() {
         verificationId: kMockVerificationId,
         smsCode: kMockSmsCode,
       );
-      await auth.signInWithCredential(credential);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(log, <Matcher>[
         isMethodCall('signInWithCredential', arguments: <String, dynamic>{
           'app': auth.app.name,
@@ -515,7 +553,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -540,7 +580,9 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -565,7 +607,9 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -590,7 +634,9 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -615,7 +661,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -638,9 +686,9 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -669,9 +717,9 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -699,8 +747,8 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -723,9 +771,9 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -755,8 +803,8 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -779,9 +827,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -809,8 +857,8 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -833,9 +881,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -1129,9 +1177,9 @@ void main() {
     });
 
     test('signInWithCustomToken', () async {
-      final FirebaseUser user =
+      final AuthResult result =
           await auth.signInWithCustomToken(token: kMockCustomToken);
-      verifyUser(user);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -1170,7 +1218,7 @@ void main() {
       await Future<void>.delayed(const Duration(seconds: 0));
 
       await simulateEvent(null);
-      await simulateEvent(mockFirebaseUser());
+      await simulateEvent(kMockUser);
 
       final FirebaseUser user1 = await events.remove();
       expect(user1, isNull);
@@ -1217,28 +1265,6 @@ void main() {
     });
   });
 }
-
-Map<String, dynamic> mockFirebaseUser(
-        {String providerId = kMockProviderId,
-        String uid = kMockUid,
-        String displayName = kMockDisplayName,
-        String photoUrl = kMockPhotoUrl,
-        String email = kMockEmail}) =>
-    <String, dynamic>{
-      'isAnonymous': true,
-      'isEmailVerified': false,
-      'providerData': <Map<String, String>>[
-        <String, String>{
-          'providerId': providerId,
-          'uid': uid,
-          'displayName': displayName,
-          'photoUrl': photoUrl,
-          'email': email,
-        },
-      ],
-      'creationTimestamp': kMockCreationTimestamp,
-      'lastSignInTimestamp': kMockLastSignInTimestamp,
-    };
 
 /// Queue whose remove operation is asynchronous, awaiting a corresponding add.
 class AsyncQueue<T> {